Move translations to currencies.ts
