+++ packages/utils/src/currencies.ts
@@ -4,8 +4,6 @@ export enum CURRENCY_ENUM {
 }
 
 interface Currency {
-  singleUnit: string
-  pluralUnit: string
   symbol: string
   code: string
 }
@@ -14,14 +12,10 @@ type CurrencyObject = { [key in CURRENCY_ENUM]: Currency }
 
 export const CURRENCIES: CurrencyObject = {
   [CURRENCY_ENUM.GOLD]: {
-    singleUnit: 'Celo gold',
-    pluralUnit: 'Celo gold',
     symbol: '',
     code: 'cGLD',
   },
   [CURRENCY_ENUM.DOLLAR]: {
-    singleUnit: 'Celo dollar',
-    pluralUnit: 'Celo dollars',
     symbol: '$',
     code: 'cUSD',
   },
@@ -46,4 +40,23 @@ export enum SHORT_CURRENCIES {
 export const currencyToShortMap = {
   [CURRENCY_ENUM.DOLLAR]: SHORT_CURRENCIES.DOLLAR,
   [CURRENCY_ENUM.GOLD]: SHORT_CURRENCIES.GOLD,
+}
+
+const currencyTranslationsENUS: { [key: string]: string } = {
+  CeloDollars: 'Celo Dollars',
+  CeloDollar: 'Celo Dollar',
+  Dollar: 'Dollar',
+  Dollars: 'Dollars',
+  dollars: 'dollars',
+}
+const currencyTranslationsESAR: { [key: string]: string } = {
+  CeloDollars: 'Celo Dólares',
+  CeloDollar: 'Celo Dólar',
+  Dollar: 'Dólar',
+  Dollars: 'Dólares',
+  dollars: 'dólares',
+}
+export const currencyTranslations: { [key: string]: any } = {
+  ['en-US']: currencyTranslationsENUS,
+  ['es-AR']: currencyTranslationsESAR,
 }
